fixed errors in persistence classes

+++ Library-Backend/src/main/java/ca/mcgill/ecse321/library/dao/LibrarianRepository.java
@@ -8,6 +8,6 @@ import ca.mcgill.ecse321.library.model.Librarian;
 @Repository
 public interface LibrarianRepository extends CrudRepository<Librarian, Integer> {
 	
-	Librarian findLibrarianbyUserid(int userId);
+	Librarian findLibrarianbyUserid(Long userId);
 
 }
